feat: don't make unnecessary requests to get buckets
CloudStorage's `check_bucket` implementation would (as the name implies) check that the bucket exists. This, however, requires a `storage.buckets.get` permission that seems to be only found by default in legacy roles.

This commit refactors so that calls to CloudStorage that don't set `auto_create_bucket` will skip the bucket check altogether